Move execution logic to its own function
The block of code which is meant to directly execute when the file is
run has now been moved to its own function, and that function is run
in a standard name == main block.  This should allow later refactors
of the code, including potentially putting the run-bot call into its
own file entirely.